[Util] Move Attribute based pipeline to Util dialect.
The initial implementation of attribute-based pipelines was restricted
to preprocessing. This is now generalized to allow applying an
attribute based pipeline specification anywhere in the compiler.

The `util.pipelines` attribute is a `DictionaryAttr` that contains a
list of pipelines that can be applied anywhere. For example, currently
a `preprocessing` pipeline can be specified as follows,

```
util.pipelines = {
    "preprocessing" : #util.pipeline<"iree-preprocessing-make-single-dispatch">
}
```

Currently `iree-compile` only understands the "preprocessing" entry,
but more such entry points can be added in the future.
