+++ compiler/src/iree/compiler/Dialect/Util/Transforms/AttrBasedPipeline.cpp
@@ -4,33 +4,33 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/compiler/Preprocessing/Common/Passes.h"
+#include "iree/compiler/Dialect/Util/Transforms/Passes.h"
 
-#include "iree/compiler/Dialect/Flow/IR/FlowDialect.h"
 #include "iree/compiler/Dialect/Util/IR/UtilDialect.h"
 #include "iree/compiler/Dialect/Util/IR/UtilOps.h"
 #include "iree/compiler/Dialect/Util/IR/UtilTypes.h"
 #include "llvm/Support/Debug.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/IR/BuiltinAttributes.h"
 #include "mlir/Interfaces/FunctionInterfaces.h"
 
 #define DEBUG_TYPE "iree-attr-based-pipeline"
 
-namespace mlir::iree_compiler::Preprocessing {
+namespace mlir::iree_compiler::IREE::Util {
 
 #define GEN_PASS_DEF_ATTRBASEDPIPELINEPASS
-#include "iree/compiler/Preprocessing/Common/Passes.h.inc" // IWYU pragma: export
+#include "iree/compiler/Dialect/Util/Transforms/Passes.h.inc" // IWYU pragma: export
 
 namespace {
 
 struct AttrBasedPipelinePass
-    : public iree_compiler::Preprocessing::impl::AttrBasedPipelinePassBase<
-          AttrBasedPipelinePass> {
+    : public impl::AttrBasedPipelinePassBase<AttrBasedPipelinePass> {
+  using Base::Base;
   void runOnOperation() override;
 };
 } // namespace
 
-static const char kPreprocessingPipelineAttrName[] = "preprocessing_pipeline";
+static const char kPipelinesAttrName[] = "util.pipelines";
 
 // Method to get the pass manager nested on a particular operation. There does
 // not seem to be a way to do this without specializing on the op itself.
@@ -52,12 +52,11 @@ void AttrBasedPipelinePass::runOnOperation() {
   });
 
   for (auto funcLikeOp : funcLikeOps) {
-    auto attr = funcLikeOp->getAttr(kPreprocessingPipelineAttrName);
+    auto attr = funcLikeOp->getAttrOfType<DictionaryAttr>(kPipelinesAttrName);
     if (!attr) {
       continue;
     }
-    auto passPipelineAttr =
-        dyn_cast<IREE::Util::PreprocessingPassPipelineAttr>(attr);
+    auto passPipelineAttr = attr.getAs<IREE::Util::PipelineAttr>(pipelineName);
     if (!passPipelineAttr) {
       funcLikeOp.emitRemark(
           "expected preprocessing_pipeline attribute to be a `StringAttr` that "
@@ -88,4 +87,4 @@ void AttrBasedPipelinePass::runOnOperation() {
   }
 }
 
-} // namespace mlir::iree_compiler::Preprocessing
+} // namespace mlir::iree_compiler::IREE::Util
