+++ compiler/src/iree/compiler/Dialect/Util/Transforms/BUILD.bazel
@@ -22,6 +22,7 @@ iree_compiler_cc_library(
     srcs = [
         "AnnotateOpOrdinals.cpp",
         "ApplyPatterns.cpp",
+        "AttrBasedPipeline.cpp",
         "CombineInitializers.cpp",
         "DropCompilerHints.cpp",
         "DumpModule.cpp",
