+++ compiler/src/iree/compiler/Dialect/Util/Transforms/Passes.h
@@ -22,6 +22,7 @@ namespace mlir::iree_compiler::IREE::Util {
 
 #define GEN_PASS_DECL_ANNOTATEOPORDINALSPASS
 #define GEN_PASS_DECL_APPLYPATTERNSPASS
+#define GEN_PASS_DECL_ATTRBASEDPIPELINEPASS
 #define GEN_PASS_DECL_COMBINEINITIALIZERSPASS
 #define GEN_PASS_DECL_DROPCOMPILERHINTSPASS
 #define GEN_PASS_DECL_DUMPMODULEPASS
