+++ compiler/src/iree/compiler/Dialect/Util/Transforms/test/BUILD.bazel
@@ -17,6 +17,7 @@ iree_lit_test_suite(
     srcs = enforce_glob(
         [
             "annotate_op_ordinals.mlir",
+            "attr_based_pipeline.mlir",
             "combine_initializers.mlir",
             "drop_compiler_hints.mlir",
             "fold_globals.mlir",
