+++ compiler/src/iree/compiler/Preprocessing/BUILD.bazel
@@ -23,6 +23,7 @@ iree_compiler_cc_library(
     deps = [
         "//compiler/src/iree/compiler/Dialect/LinalgExt/IR",
         "//compiler/src/iree/compiler/Dialect/Util/IR",
+        "//compiler/src/iree/compiler/Dialect/Util/Transforms",
         "//compiler/src/iree/compiler/DispatchCreation",
         "//compiler/src/iree/compiler/GlobalOptimization",
         "//compiler/src/iree/compiler/Pipelines:Options",
