+++ compiler/src/iree/compiler/Preprocessing/Common/BUILD.bazel
@@ -36,7 +36,6 @@ iree_compiler_cc_library(
     name = "Transforms",
     srcs = [
         "ApplyPDLPatterns.cpp",
-        "AttrBasedPipelinePass.cpp",
         "ConvertConv2DToImg2Col.cpp",
         "ConvertConvFilterToChannelsLast.cpp",
         "ConvertConvToChannelsLast.cpp",
