+++ compiler/src/iree/compiler/Preprocessing/Common/CMakeLists.txt
@@ -27,7 +27,6 @@ iree_cc_library(
     "Passes.h.inc"
   SRCS
     "ApplyPDLPatterns.cpp"
-    "AttrBasedPipelinePass.cpp"
     "ConvertConv2DToImg2Col.cpp"
     "ConvertConvFilterToChannelsLast.cpp"
     "ConvertConvToChannelsLast.cpp"
