+++ compiler/src/iree/compiler/Preprocessing/Common/Passes.td
@@ -27,19 +27,6 @@ def ApplyPDLPatternsPass : Pass<"iree-preprocessing-apply-pdl-patterns", "Module
   ];
 }
 
-def AttrBasedPipelinePass : Pass<"iree-preprocessing-attr-based-pipeline", ""> {
-  let summary = "Run a pass pipeline specified as an attribute on any function-like operation";
-  let description = [{
-    The textual representation of the pass pipeline specified in the attribute
-    `preprocessing_pipeline=#util.preprocessing_pipeline<...>` is to be run on the
-    function-like operation.
-  }];
-  let dependentDialects = [
-    "iree_compiler::IREE::Flow::FlowDialect",
-    "iree_compiler::IREE::Util::UtilDialect",
-  ];
-}
-
 def ConvertConv2DToImg2ColPass :
     Pass<"iree-preprocessing-convert-conv2d-to-img2col", ""> {
   let summary = "Convert linalg convolution ops to matmul img2col based implementation";
