+++ compiler/src/iree/compiler/Preprocessing/Common/test/BUILD.bazel
@@ -16,7 +16,6 @@ iree_lit_test_suite(
     name = "lit",
     srcs = enforce_glob(
         [
-            "attr_based_pipeline.mlir",
             "conv2d_to_img2col.mlir",
             "conv_filter_to_channels_last.mlir",
             "conv_to_channels_last.mlir",
