+++ compiler/src/iree/compiler/Preprocessing/Common/test/CMakeLists.txt
@@ -14,7 +14,6 @@ iree_lit_test_suite(
   NAME
     lit
   SRCS
-    "attr_based_pipeline.mlir"
     "conv2d_to_img2col.mlir"
     "conv_filter_to_channels_last.mlir"
     "conv_to_channels_last.mlir"
